# cmd/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Noble blockchain indexer",
        "title": "Noble Indexer API",
        "termsOfService": "https://bakingbad.dev/terms",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/baking-bad/noble-indexer",
            "email": "hello@bakingbad.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/baking-bad/noble-indexer/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "noble.dipdup.net",
    "basePath": "/v1",
    "paths": {
        "/address": {
            "get": {
                "description": "Returns a paginated list of addresses with their balances. Can be filtered to show only contract addresses.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "List addresses",
                "operationId": "list-address",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of addresses to return (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of addresses to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "Sort order (default: asc)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "value",
                            "first_height",
                            "last_height"
                        ],
                        "type": "string",
                        "description": "Field to sort by (default: id)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "If true, return only addresses that are smart contracts (default: false)",
                        "name": "only_contracts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of addresses with their balances",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Address"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/address/{hash}": {
            "get": {
                "description": "Returns detailed information about a specific address including its balance, contract status, and activity history",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "Get address by hash",
                "operationId": "get-address",
                "parameters": [
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "description": "Address hash in hexadecimal format (e.g., 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb)",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address information with balance",
                        "schema": {
                            "$ref": "#/definitions/responses.Address"
                        }
                    },
                    "204": {
                        "description": "Address not found"
                    },
                    "400": {
                        "description": "Invalid address hash format",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.Address": {
            "description": "Noble address information",
            "type": "object",
            "properties": {
                "balance": {
                    "$ref": "#/definitions/responses.Balance"
                },
                "deployed_contracts": {
                    "type": "integer",
                    "example": 7
                },
                "first_height": {
                    "type": "integer",
                    "example": 100
                },
                "hash": {
                    "type": "string",
                    "example": "0xd90d69b7cf347b5bfe0719baf7eef310c085e46b"
                },
                "id": {
                    "type": "integer",
                    "example": 321
                },
                "interactions": {
                    "type": "integer",
                    "example": 890
                },
                "is_contract": {
                    "type": "boolean",
                    "example": false
                },
                "last_height": {
                    "type": "integer",
                    "example": 100
                },
                "txs_count": {
                    "type": "integer",
                    "example": 23456
                }
            }
        },
        "responses.Balance": {
            "description": "Balance of address information",
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string",
                    "example": "noble"
                },
                "value": {
                    "type": "string",
                    "example": "10000000000"
                }
            }
        },
...
    }
}